+++ backend/application/api/src/main/kotlin/io/raemian/api/comment/CommentService.kt
@@ -1,5 +1,7 @@
 package io.raemian.api.comment
 
+import io.raemian.api.comment.controller.request.WriteCommentRequest
+import io.raemian.api.comment.controller.response.CommentsResponse
 import io.raemian.api.comment.event.UpdateLastCommentReadAtEvent
 import io.raemian.api.support.error.CoreApiException
 import io.raemian.api.support.error.ErrorInfo
