+++ backend/application/api/src/main/kotlin/io/raemian/api/comment/controller/request/WriteCommentRequest.kt
@@ -1,4 +1,4 @@
-package io.raemian.api.comment
+package io.raemian.api.comment.controller.request
 
 data class WriteCommentRequest(
     val content: String,
