+++ backend/application/api/src/main/kotlin/io/raemian/api/comment/controller/response/CommentsResponse.kt
@@ -1,4 +1,4 @@
-package io.raemian.api.comment
+package io.raemian.api.comment.controller.response
 
 import io.raemian.storage.db.core.comment.Comment
 import io.raemian.storage.db.core.user.User
